Add script to trigger direct archiving
This is the 1st iteration of this script. It only needs the submission UUID as a parameter. You can configure the environment you would like to run in the `ENVIRONMENT` variable in the script or you could set it as an environment variable.

Known issue: If a study has been already submitted to ENA you can not submit it again. ENA sends an error response. The script is currently not processing that error response. I am going to create a ticket for that job.